sync latest changes from Meatloaf

+++ lib/meatloaf/file/p00.cpp
@@ -6,8 +6,8 @@
  * Streams
  ********************************************************/
 
-uint16_t P00MStream::readFile(uint8_t* buf, uint16_t size) {
-    uint16_t bytesRead = 0;
+uint32_t P00MStream::readFile(uint8_t* buf, uint32_t size) {
+    uint32_t bytesRead = 0;
 
     bytesRead += containerStream->read(buf, size);
     _position += bytesRead;
